# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gin-admin-api 后台管理接口文档",
        "title": "gin-admin-api",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/post": {
            "post": {
                "description": "创建岗位，岗位编码全局唯一",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "岗位中心"
                ],
                "summary": "创建岗位",
                "parameters": [
                    {
                        "description": "创建岗位参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gin-admin-api_internal_api_post_dto.CreateSysPostDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/post/detail/{id}": {
            "get": {
                "description": "根据岗位 ID 获取岗位详情，自动应用数据权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "岗位中心"
                ],
                "summary": "获取岗位详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "岗位ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "岗位详情",
                        "schema": {
                            "$ref": "#/definitions/gin-admin-api_internal_api_post_vo.SysPostVO"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/post/list": {
            "get": {
                "description": "根据名称、编码和状态获取岗位列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "岗位中心"
                ],
                "summary": "获取岗位列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "岗位名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "岗位编码",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态：1正常，2禁用，0全部",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "岗位列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/gin-admin-api_internal_api_post_vo.SysPostVO"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/post/page": {
            "post": {
                "description": "根据岗位名称、编码、状态分页查询岗位列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "岗位中心"
                ],
                "summary": "分页获取岗位",
                "parameters": [
                    {
                        "description": "岗位分页查询参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SysPostPageDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "岗位分页列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/gin-admin-api_internal_api_post_vo.SysPostVO"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/post/{id}": {
            "put": {
                "description": "根据岗位 ID 修改岗位基础信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "岗位中心"
                ],
                "summary": "修改岗位",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "岗位ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改岗位参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gin-admin-api_internal_api_post_dto.ModifySysPostDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据岗位 ID 删除岗位；已绑定账号时不能删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "岗位中心"
                ],
                "summary": "删除岗位",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "岗位ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.SysPostPageDTO": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "岗位编码，模糊查询",
                    "type": "string"
                },
                "name": {
                    "description": "岗位名称，模糊查询",
                    "type": "string"
                },
                "pageNumber": {
                    "description": "页码",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：1正常，2禁用；0表示全部",
                    "type": "integer"
                }
            }
        },
...
        "gin-admin-api_internal_api_post_dto.CreateSysPostDTO": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "code": {
                    "description": "岗位编码，全局唯一",
                    "type": "string",
                    "maxLength": 60
                },
                "name": {
                    "description": "岗位名称",
                    "type": "string",
                    "maxLength": 50
                },
                "remark": {
                    "description": "备注",
                    "type": "string",
                    "maxLength": 255
                },
                "sort": {
                    "description": "排序，值越小越靠前",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：1正常，2禁用；不传默认1",
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                }
            }
        },
        "gin-admin-api_internal_api_post_dto.ModifySysPostDTO": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "code": {
                    "description": "岗位编码，全局唯一",
                    "type": "string",
                    "maxLength": 60
                },
                "name": {
                    "description": "岗位名称",
                    "type": "string",
                    "maxLength": 50
                },
                "remark": {
                    "description": "备注",
                    "type": "string",
                    "maxLength": 255
                },
                "sort": {
                    "description": "排序，值越小越靠前",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：1正常，2禁用；不传默认1",
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                }
            }
        },
        "gin-admin-api_internal_api_post_vo.SysPostVO": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "岗位编码，全局唯一",
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间，Unix秒",
                    "type": "integer"
                },
                "createdBy": {
                    "description": "创建人账号id",
                    "type": "integer"
                },
                "id": {
                    "description": "主键id",
                    "type": "integer"
                },
                "name": {
                    "description": "岗位名称",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "sort": {
                    "description": "排序，值越小越靠前",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：1正常，2禁用",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "更新时间，Unix秒",
                    "type": "integer"
                },
                "updatedBy": {
                    "description": "更新人账号id",
                    "type": "integer"
                }
            }
        },
...
    }
}